feat(ai): add chatbot assistant

+++ src/main/java/org/hlopes/library/dto/AddToLibraryRequest.java
@@ -2,7 +2,7 @@ package org.hlopes.library.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record AddToLibraryRequest(@NotNull Long externalId, @NotNull String mediaType, String status, Integer rating) {
+public record AddToLibraryRequest(@NotNull Long externalId, @NotNull String mediaType, String status, Short rating) {
     public AddToLibraryRequest(@NotNull Long externalId, @NotNull String mediaType) {
         this(externalId, mediaType, null, null);
     }
